Add support for build scripts in the config directory
Currently Dyre does not support build scripts. A build script allows
the config project to be built in any special way that a user may
desire. It is also a feature of XMonad's config system, which Dyre
should parity.

A build script can now be optionally used. This script takes one
argument, the path to copy the built executable to, which Dyre gives
when calling the script. Dyre will prioritize the build script over a
source file.